Connecting backend to frontend via api
I wrote a simple api script to connect the backend to the front end using this video tutorial. We will likely have to make it more robust for the project, but hopefully it is enough to get us started. https://www.youtube.com/watch?v=w3vs4a03y3I